data: interactor: Move Save to a specific file
As part of the interactor refactor efforts, this patch moves the `Save`
operation to a specific file. In addition, considering the
`Authenticate` is used by most of the operations, it's moved to near
`DataInteractor` structure and its interface, which will be implemented
by all the operations.
